+++ cuda_src/setup.cu
@@ -415,7 +415,7 @@ void CUDAPathTracer::buildBVH()
     printf("cudaMalloc\n");
     cudaMalloc((void**)&gpu_leafNodes, numObjects * sizeof(GPUBVHNode));
     cudaMalloc((void**)&gpu_internalNodes, (numObjects - 1) * sizeof(GPUBVHNode));
-    cudaMalloc((void**)&gpu_sortedMortonCodes, numObjects * sizeof(unsigned int));
+    cudaMalloc((void**)&gpu_sortedMortonCodes, numObjects * sizeof(unsigned long long));
     cudaMalloc((void**)&BVHPrimMap, numObjects * sizeof(int));
 
     // gpu_sortedMortonCodes = sortedMortonCodes;
@@ -468,7 +468,7 @@ void CUDAPathTracer::buildBVH()
     // unsigned int* keys = thrust::raw_pointer_cast(const_bvhparams.sortedMortonCodes);
     // int* data = thrust::raw_pointer_cast(const_bvhparams.sortedObjectIDs);
     printf("thrustSort\n");
-    thrust::device_ptr<unsigned int> keys = thrust::device_pointer_cast(gpu_sortedMortonCodes);
+    thrust::device_ptr<unsigned long long> keys = thrust::device_pointer_cast(gpu_sortedMortonCodes);
     thrust::device_ptr<int> data = thrust::device_pointer_cast(BVHPrimMap);
     thrust::sort_by_key(keys, keys + numObjects, data);
     err = cudaPeekAtLastError();
